Move app logic to file
